requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-google-genai==1.0.3
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
torch==2.0.1
transformers==4.31.0
faiss-cpu
pypdf
google-generativeai
python-dotenv
streamlit